chore: suppress GitHub test secret payloads
Update the secret detected metric filter to omit API keys with the
work `-revoked` in them so that GitHub's tests of our secret
scanning endpoint do not trigger alarms.

Update the secret detected alarm to remove the `OK` action to
remove channel noise.

Update the log insight query names to include the product name
since this service is in a shared account.